# docs-site/docs/api/openapi/core-full.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AstroEngine API",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/scan/progressions": {
      "post": {
        "tags": [
          "scan"
        ],
        "summary": "Api Scan Progressions",
        "operationId": "api_scan_progressions_v1_scan_progressions_post",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "description": "Maximum number of hits to return (default 500, maximum 2000).",
              "default": 500,
              "title": "Limit"
            },
            "description": "Maximum number of hits to return (default 500, maximum 2000)."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of hits to skip from the start of the result set.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of hits to skip from the start of the result set."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Scan request payload",
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scan/directions": {
      "post": {
        "tags": [
          "scan"
        ],
        "summary": "Api Scan Directions",
        "operationId": "api_scan_directions_v1_scan_directions_post",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "description": "Maximum number of hits to return (default 500, maximum 2000).",
              "default": 500,
              "title": "Limit"
            },
            "description": "Maximum number of hits to return (default 500, maximum 2000)."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of hits to skip from the start of the result set.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of hits to skip from the start of the result set."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Scan request payload",
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scan/transits": {
      "post": {
        "tags": [
          "scan"
        ],
        "summary": "Api Scan Transits",
        "operationId": "api_scan_transits_v1_scan_transits_post",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "description": "Maximum number of hits to return (default 500, maximum 2000).",
              "default": 500,
              "title": "Limit"
            },
            "description": "Maximum number of hits to return (default 500, maximum 2000)."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of hits to skip from the start of the result set.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of hits to skip from the start of the result set."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Scan request payload",
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scan/returns": {
      "post": {
        "tags": [
          "scan"
        ],
        "summary": "Api Scan Returns",
        "operationId": "api_scan_returns_v1_scan_returns_post",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "description": "Maximum number of hits to return (default 500, maximum 2000).",
              "default": 500,
              "title": "Limit"
            },
            "description": "Maximum number of hits to return (default 500, maximum 2000)."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of hits to skip from the start of the result set.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of hits to skip from the start of the result set."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Scan request payload",
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Hit": {
        "properties": {
          "ts": {
            "type": "string",
            "title": "Ts"
          },
          "moving": {
            "type": "string",
            "title": "Moving"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "aspect": {
            "type": "integer",
            "title": "Aspect"
          },
          "orb": {
            "type": "number",
            "title": "Orb"
          },
          "orb_allow": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orb Allow"
          },
          "motion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motion"
          },
          "family": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family"
          },
          "lon_moving": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon Moving"
          },
          "lon_target": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon Target"
          },
          "delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta"
          },
          "offset": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset"
          },
          "speed_deg_per_day": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Deg Per Day"
          },
          "retrograde": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrograde"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "ts",
          "moving",
          "target",
          "aspect",
          "orb"
        ],
        "title": "Hit"
      },
...
      "ScanResponse": {
        "properties": {
          "method": {
            "type": "string",
            "title": "Method"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/Hit"
            },
            "type": "array",
            "title": "Hits"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "export": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Export"
          }
        },
        "type": "object",
        "required": [
          "method",
          "hits",
          "count"
        ],
        "title": "ScanResponse",
        "description": "count reflects total hits before pagination is applied"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}